refactor: :fire: Rename form validation function to camelCase, Delete LoginButton component
